+++ backend/src/main/java/fr/imoblex/modules/property/dto/PropertyCreateRequest.java
@@ -31,8 +31,8 @@ public class PropertyCreateRequest {
     private String postalCode;
     private String department;
     private String region;
-    private Double latitude;
-    private Double longitude;
+    private BigDecimal latitude;
+    private BigDecimal longitude;
     private Boolean addressVisible = true;
 
     // Prix
@@ -48,9 +48,9 @@ public class PropertyCreateRequest {
     private String agencyFeesInfo;
 
     // Surfaces & pièces
-    private Double livingArea;
-    private Double totalArea;
-    private Double landArea;
+    private BigDecimal livingArea;
+    private BigDecimal totalArea;
+    private BigDecimal landArea;
     private Integer rooms;
     private Integer bedrooms;
     private Integer bathrooms;
@@ -84,8 +84,8 @@ public class PropertyCreateRequest {
     // DPE
     private String dpeClass;
     private String gesClass;
-    private Double dpeValue;
-    private Double gesValue;
+    private BigDecimal dpeValue;
+    private BigDecimal gesValue;
     private LocalDate dpeDoneDate;
     private Boolean dpeExempt;
